requirements.txt
tqdm
sqlite3
requests
python-dotenv
OTXv2 # for AlienVault
